requirements.txt
requests==2.24.0
uvicorn==0.11.8
pytest==5.3.5
fastapi==0.61.1
PyYAML==5.3.1
python-multipart==0.0.5
